REBUILD all files using latest `dev' upstream
These new files contain the latest refinements in upstream Hack, most
notable of which is the proper alignment of delimiters to the baseline.

The underlying intention of my patches is to draw a more clear
distinction between roman and italic styles and to use glyphs that are
more appropriate for a workhorse typeface.

Hack is sublime.  These patches just add the finishing touches (to my
eyes, of course).